ai/requirements.txt
# Core dependencies - Python 3.13 compatible
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
numpy>=1.24.0
Pillow>=10.0.0

# Optional AI dependencies (install manually if needed)
# ultralytics  # For YOLO object detection
# opencv-python  # For image processing
# pytesseract  # For OCR
# torch  # For deep learning models